+++ changelog.md
@@ -1,15 +1,17 @@
 **CHANGELOG :**
 
-+ Version 0.18.1.1 : Initial code
 
-        - getting data from Cryptonator
-        - getting data from CryptoCompare
-        - replying by quote to the user (on /convert)
-        - emoji :rocket: on results
-        - inline requests supported
-        - inline error showing
-        - about with emojis
++ Version 0.18.1.3
 
+        - le bot dit bonjour/au revoir
+        - ajout de /help et de /about
+        - la langue par défaut l'anglais
+        - ajout de plus d'explication dans /help
+        - affichage du prix en BTC dans /snap
+                [TODO: formater en SAT]
+        - close_file ajouté aux opérations I/O
+        
+        
 + Version 0.18.1.2
 
         - ajout de l'API coinmarketcap
@@ -20,3 +22,18 @@
         - affichage du résultat avec des émojis
         - support d'un cache pour la liste des monnaies
         - ajout de /ticker
+        
+
++ Version 0.18.1.1 : Initial code
+
+        - getting data from Cryptonator
+        - getting data from CryptoCompare
+        - replying by quote to the user (on /convert)
+        - emoji :rocket: on results
+        - inline requests supported
+        - inline error showing
+        - about with emojis
+
+
+        
+